Fixes #2361 - Move DELETE related and DELETE + PUT tests to common class

+++ catalog-rest-service/src/main/java/org/openmetadata/catalog/resources/databases/DatabaseResource.java
@@ -408,6 +408,7 @@ public class DatabaseResource {
           boolean recursive,
       @PathParam("id") String id)
       throws IOException {
+    SecurityUtil.checkAdminOrBotRole(authorizer, securityContext);
     DeleteResponse<Database> response = dao.delete(securityContext.getUserPrincipal().getName(), id, recursive);
     return response.toResponse();
   }
